start work on criticality
